requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
